view models refactored
